requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
mysqlclient==2.1.1
django-decouple==2.1
djangorestframework-simplejwt==5.5.0
django-filter==25.1
